requirements.txt
PyQt5==5.15.9
Pillow==10.0.0
pywin32==306
pyinstaller==6.3.0 